backend2/requirements.txt
python-dotenv
groq
Flask
flask_cors
AppOpener
pywhatkit
bs4
# pillow
rich
# requests
Keyboard
cohere
googlesearch-python
# selenium
# mtranslate
# pygame
# edge-tts
# PyQt5
webdriver-manager
